ailab-llama-search/requirements.txt
llama-index
llama-index-vector-stores-postgres
llama-index-embeddings-azure-openai
llama-index-storage-kvstore-postgres
python-dotenv
